43 change scheduler to be setup correctly
